requirements.txt
aiohttp>=3.9.0
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
distro
dnspython
email-validator
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
frozenlist
greenlet>=3.0.3
gunicorn
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mdurl
midjourney_api
multidict
openai>=1.0.0
orjson
packaging
pydantic>=2.5.0
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
requests
rich
shellingham
sniffio
SQLAlchemy>=2.0.0
starlette
tqdm
typer
typing_extensions
ujson
urllib3
uvicorn
uvloop>=0.21.0; sys_platform != "win32"
watchfiles
websockets
Werkzeug
yarl
